requirements.txt
flask==3.0.3
google-generativeai==0.8.3
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.20.0
gunicorn==22.0.0
python-dotenv==1.0.1
cachetools==5.3.3
Flask-Limiter==3.8.0
pytest==8.2.2
pytest-flask==1.3.0
